requirements.txt
discord.py-self
discord.py
colorama
os
io
python-dotenv
rich
sqlite3
aiohttp